채팅방 업데이트
1. 나가기 기능 업그레이드
	기존에 나갔을 경우 사용자가 얼마나 남았는지와는 상관 없이
	데이터를 전부 삭제했었음. 오늘 작업을 통해서 채팅방
	데이터베이스에 membersNum필드를 추가하여 채팅방에 인원이 얼마나
	남았는지 표시 + 그것을 이용해서 채팅방에 아무도 없는 조건이
	되었을 경우 해당 문서를 삭제하게끔 만들었음. 문제사항으로
	사용자가 0명이 되어 members가 0이 된다면 getMembersGroup에서
	Null인 List를 넘겨주기에 에러가 발생한다는 것...

	또한 해당문서를 삭제할 때에 Firebase Storage에 저장된 이미지 파일도
	삭제하게끔 만들었음. Firebase Storage에서 이미지 저장을 채팅방
	ID 를 이름으로 가진 폴더에다가 이미지를 저장했는데, 그 폴더를
	삭제하는 것이 불가능했다. For-each를 이용해서 해당 폴더 내의
	파일을 일일이 삭제해야함.

2. GroupTile 업데이트
	새로운 채팅방을 만들었을 때에 recentMessage에 null값이 들어가
	오류가 났다. 하지만 try-catch를 이용해서 수정, 오류를 해결했다.

	또한 사진을 보냈을 경우에 Download URL이 GroupTile에 표시가
	되어 Overflow가 일어났으나, recentMessageType이라는 필드값을
	채팅방 doc에 추가함으로써 메시지인지 이미지인지를 구분하도록
	하였고, 그것을 토대로 이미지일 경우에는 '사진'이라고 뜨게끔
	만들었음.

3. 사용자 목록 표시 업데이트
	채팅방 내부에서 사용자의 목록을 Drawer에서 보여주는데,
	Expanded를 쓰다보니 나가기 버튼이 사용자의 수와 상관없이
	아래쪽에 가 있었다. 따라서 getGroupMembers 위젯에서
	shrinkWrap속성을 True로 주었더니 알맞게 표시가 되었다.

+++ lib/services/database_service.dart
@@ -1,4 +1,6 @@
+
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_storage/firebase_storage.dart';
 
 class DatabaseService {
 
@@ -31,6 +33,7 @@ class DatabaseService {
       'groupIcon': '',
       'admin': userName,
       'members': [],
+      'membersNum' : 1,
       //'messages': ,
       'groupId': '',
       'recentMessage': '',
@@ -56,9 +59,11 @@ class DatabaseService {
     DocumentSnapshot userDocSnapshot = await userDocRef.get();
 
     DocumentReference groupDocRef = groupCollection.doc(groupId);
+    DocumentSnapshot groupDocSnapshot = await groupDocRef.get();
+    FirebaseStorage desertRef =  FirebaseStorage.instance;
 
+    int membersNum = await groupDocSnapshot.data()['membersNum'];
     List<dynamic> groups = await userDocSnapshot.data()['groups'];
-
     if(groups.contains(groupId + '_' + groupName)) {
       //print('hey');
       await userDocRef.update({
@@ -66,10 +71,17 @@ class DatabaseService {
       });
 
       await groupDocRef.update({
-        'members': FieldValue.arrayRemove([uid + '_' + userName])
+        'members': FieldValue.arrayRemove([uid + '_' + userName]),
+        'membersNum' : FieldValue.increment(-1)
       });
-
-      groupDocRef.delete();
+      if(membersNum == 1){
+        desertRef.ref().child(groupId+'/').listAll().then((value){
+          value.items.forEach((element) {
+            element.delete();
+          });
+        });
+         groupDocRef.delete();
+      }
     }
     else {
       //print('nay');
@@ -78,7 +90,8 @@ class DatabaseService {
       });
 
       await groupDocRef.update({
-        'members': FieldValue.arrayUnion([uid + '_' + userName])
+        'members': FieldValue.arrayUnion([uid + '_' + userName]),
+        'membersNum' : FieldValue.increment(1)
       });
     }
   }
@@ -121,12 +134,13 @@ class DatabaseService {
 
 
   // send message
-  sendMessage(String groupId, chatMessageData) {
+  sendMessage(String groupId, chatMessageData, String type) {
     FirebaseFirestore.instance.collection('groups').doc(groupId).collection('messages').add(chatMessageData);
     FirebaseFirestore.instance.collection('groups').doc(groupId).update({
       'recentMessage': chatMessageData['message'],
       'recentMessageSender': chatMessageData['sender'],
       'recentMessageTime': chatMessageData['time'],
+      'recentMessageType' : chatMessageData['type']
     });
   }
